added some jsdocs
