+++ packages/common/src/strings.ts
@@ -1,14 +1,25 @@
 import { getIn } from "./objects";
 
+/**
+ * Throws if value is not a string
+ * @param value 
+ * @param errorMessage 
+ */
 export function assertIsString(value: any, errorMessage = 'Value is not string'): asserts value is string {
     if (typeof value !== 'string') {
         throw Error(errorMessage);
     }
 }
 
-// https://drafts.csswg.org/cssom/#escape-a-character-as-code-point
-export const escapeCSS = (s: string) =>
-    s.replace(/\W/giu, (char) => {
+// 
+/**
+ * Replaced non alphanumeric character with CSS unicode representation
+ * @see https://drafts.csswg.org/cssom/#escape-a-character-as-code-point
+ * @param str 
+ * @returns CSS safe string
+ */
+export const escapeCSS = (str: string) =>
+    str.replace(/\W/giu, (char) => {
         const code = char.codePointAt(0) ?? 0xfffd;
         return `\\${code.toString(16)} `;
     });
@@ -20,14 +31,34 @@ export enum NamingConvention {
     CamelCase = 'camel-case',
 }
 
+/**
+ * 
+ * @param namingConvention 
+ * @returns true if namingConvention is a supported NamingConvention
+ */
 export function isValidNamingConvention(namingConvention: string): namingConvention is NamingConvention {
     return Object.values(NamingConvention).some((value) => value === namingConvention);
 }
 
+/**
+ * Capitalize the first letter of a string
+ * @param val 
+ * @returns val with the first letter capitalized 
+ */
 export function capitalizeFirstLetter(val: string): string {
     return val.length === 0 ? val : val.charAt(0).toUpperCase() + val.slice(1);
 }
 
+/**
+ * Breaks down a string to words, dropping non letters and numbers
+ * @example <caption>Spaces</caption> splitIntoWords("Hello world") => ["Hello", "world"]
+ * @example <caption>Numbers</caption>  splitIntoWords("Hello123world") => ["Hello", "123" "world"]
+ * @example <caption>Acronyms</caption> splitIntoWords("Hello WRL") => ["Hello", "WRL"]
+ * @example <caption>Capitalized</caption> splitIntoWords("HelloWorld") => ["Hello", "World"]
+ * @example <caption>Others characters</caption> splitIntoWords("Hello_world--"") => ["Hello", "world"]
+ * @param str 
+ * @returns An array of words contained in str
+ */
 export const splitIntoWords = (str: string): string[] => {
     let words = str.match(/[a-z0-9]+/gi) ?? [];
     words = words.flatMap((w) => w.split(/(\d+)/g)); // Numbers
@@ -36,20 +67,42 @@ export const splitIntoWords = (str: string): string[] => {
     return words.filter((w) => w);
 };
 
+/**
+ * Converts a string to kebab-case
+ * @param str 
+ * @returns str in kebab-case-convention
+ */
 export function toKebabCase(str: string): string {
     return splitIntoWords(str).join('-').toLowerCase();
 }
 
+/**
+ * Converts a string to PascalCase
+ * @param str 
+ * @returns str in PascalCaseConvention
+ */
 export function toPascalCase(str: string): string {
     const words = splitIntoWords(str).map((word) => capitalizeFirstLetter(word.toLowerCase()));
     return words.join('');
 }
 
+/**
+ * Similar to toPascalCase, but drops heading non-letters
+ * @example toPascalCaseJsIdentifier("123helloWorld") => "HelloWorld"
+ * @see toPascalCase
+ * @param str 
+ * @returns str in PascalCaseConvention that starts with a letter
+ */
 export function toPascalCaseJsIdentifier(str: string): string {
     str = str.replace(/^[^a-z]+/i, ''); // must start with a letter
     return toPascalCase(str);
 }
 
+/**
+ * Converts a string to camelCase
+ * @param str 
+ * @returns str in camelCaseConvention
+ */
 export function toCamelCase(str: string): string {
     const words = splitIntoWords(str).map((word, index) =>
         index > 0 ? capitalizeFirstLetter(word.toLowerCase()) : word.toLowerCase()
@@ -57,6 +110,12 @@ export function toCamelCase(str: string): string {
     return words.join('');
 }
 
+/**
+ * Converts string formatting to a naming convention
+ * @param str 
+ * @param namingConvention 
+ * @returns 
+ */
 export function toNamingConvention(str: string, namingConvention: NamingConvention): string {
     switch (namingConvention) {
         case NamingConvention.KebabCase:
@@ -91,10 +150,10 @@ export function toCSSCamelCase(str: string): string {
 }
 
 /**
- * returns zero based line number and character
+ * Finds line an column by position index  
  * @param content
  * @param pos
- * @param newline
+ * @param newline zero based line number and character
  */
 export function indexToLineAndColumn(
     content: string,
@@ -127,6 +186,12 @@ export function indexToLineAndColumn(
     };
 }
 
+/**
+ * 
+ * @param str 
+ * @param substr 
+ * @returns true is str contains substr, ignoring capitalization
+ */
 export function includesCaseInsensitive(str: string, substr: string): boolean {
     return str.toLowerCase().includes(substr.toLowerCase());
 }
@@ -148,6 +213,12 @@ export function equalIdents(reference: string, modified: string, newline = '\n')
         .join(newline);
 }
 
+/**
+ * Remove line indentation (heading whitespace)
+ * @param modified 
+ * @param separator 
+ * @returns 
+ */
 export function noIdents(modified: string, separator = '\n') {
     const modifiedArr = modified.split(separator);
     return modifiedArr
@@ -175,24 +246,4 @@ export function templateCompilerProvider(context: Record<string, any>) {
             return valueInContext !== undefined ? valueInContext : match;
         });
     };
-}
-
-/**
- * Ensure a single heading/trailing backslash (/) of a single line string
- * @param str 
- * @param type 'heading'|'trailing'|'both'|'none'
- * @returns 
- */
-export function backSlash(str:string, type:'heading'|'trailing'|'both'|'none') {
-    const s = str.replace(/^\/+|\/+$/,'')
-    switch(type){
-        case 'both':
-            return `/${s}/`;
-        case 'trailing':
-            return `${s}/`;
-        case 'heading':
-            return `/${s}`;
-        default:
-            return s      
-    }
 }
